Move `TreeKernelPosterior` to experiments module.
