+++ tests/scripts/test_infer_graph_posterior.py
@@ -2,8 +2,7 @@ from pathlib import Path
 import pickle
 import pytest
 from scipy import stats
-from sklearn.exceptions import NotFittedError
-from summaries.scripts.infer_graph_posterior import __main__, TreeKernelPosterior
+from summaries.scripts.infer_graph_posterior import __main__
 from summaries.scripts.simulate_data import __main__ as __main__simulate_data, GraphSimulationConfig
 from unittest import mock
 
@@ -30,10 +29,3 @@ def test_infer_graph_posterior(tmp_path: Path) -> None:
     pearsonr = stats.pearsonr(observed["params"].ravel(), output["map_estimate"].ravel())
     assert pearsonr.statistic > 0.5
     assert pearsonr.pvalue < 1e-3
-
-
-def test_posteriot_not_fitted() -> None:
-    with pytest.raises(NotFittedError):
-        TreeKernelPosterior(stats.uniform(0, 1))._log_target(0.5)
-    with pytest.raises(NotFittedError):
-        TreeKernelPosterior(stats.uniform(0, 1)).log_prob(0.5)
